pipeline/src/components/cryto_ranker/src/requirements.txt
cryptowatch-sdk